update ocs 2.12.0

+++ plugins/main_sections/ms_computer/ms_computer.php
@@ -33,8 +33,9 @@ require_once('require/extensions/ExtensionHook.php');
 require_once('require/function_machine.php');
 require_once('require/function_files.php');
 require_once('ms_computer_views.php');
+require_once('require/archive/ArchiveComputer.php');
 //recherche des infos de la machine
-$item = info($protectedGet, $protectedPost['systemid']);
+$item = info($protectedGet, $protectedPost['systemid'] ?? '');
 if (!is_object($item)) {
     msg_error($item);
     require_once(FOOTER_HTML);
@@ -59,13 +60,14 @@ echo '<div class="col col-md-10">';
 if (isset($protectedPost["WOL"]) && $protectedPost["WOL"] == 'WOL' && $_SESSION['OCS']['profile']->getRestriction('WOL', 'NO') == "NO") {
     require_once('require/wol/WakeOnLan.php');
     $wol = new Wol();
-    $sql = "select MACADDR,IPADDRESS from networks WHERE (hardware_id=%s) and status='Up'";
+    $sql = "select MACADDR,IPADDRESS,IPMASK from networks WHERE (hardware_id=%s) and status='Up'";
     $arg = array($item->ID);
     $resultDetails = mysql2_query_secure($sql, $_SESSION['OCS']["readServer"], $arg);
     $msg = "";
 
     while ($wol_item = mysqli_fetch_object($resultDetails)) {
-        $wol->look_config_wol($wol_item->IPADDRESS, $wol_item->MACADDR);
+        $broadcast = long2ip(ip2long($wol_item->IPADDRESS) | ~ip2long($wol_item->IPMASK));
+        $wol->look_config_wol($broadcast, $wol_item->MACADDR);
 
         if ($wol->wol_send == $l->g(1282)) {
             msg_info($wol->wol_send . "=>" . $wol_item->MACADDR . "/" . $wol_item->IPADDRESS);
@@ -76,6 +78,16 @@ if (isset($protectedPost["WOL"]) && $protectedPost["WOL"] == 'WOL' && $_SESSION[
 }
 
 show_computer_title($item);
+
+
+$archive = new ArchiveComputer();
+if (isset($protectedPost["ARCHIVE"]) && $protectedPost['ARCHIVE'] == $l->g(1551)) {
+    $archive->archive($item->ID);
+    unset($protectedPost['ARCHIVE']);
+} elseif (isset($protectedPost["ARCHIVE"]) && $protectedPost['ARCHIVE'] ==  $l->g(1552)) {
+    $archive->restore($item->ID);
+    unset($protectedPost['ARCHIVE']);
+}
 
 show_computer_actions($item);
 
@@ -123,14 +135,14 @@ if (isset($protectedGet['cat']) && in_array($protectedGet['cat'], array('softwar
     }
 } else if (isset($protectedGet['option'])) {
     // If specific plugin
-    $plugin = $plugins[$protectedGet['option']];
+    $plugin = $plugins[$protectedGet['option']] ?? '';
     if($plugin != null){
         $plugin_file = PLUGINS_DIR . "computer_detail/" . $plugin->getId() . "/" . $plugin->getId() . ".php";
     }else{
         $file_extension = EXT_DL_DIR . $protectedGet['option'] . "/cd_" . $protectedGet['option'] . "/cd_" . $protectedGet['option'] .".php";
     }
 
-    if (file_exists($plugin_file) || file_exists($file_extension)) {
+    if ((isset($plugin_file) && (file_exists($plugin_file)) || file_exists($file_extension))) {
         if (!AJAX) {
             if(file_exists($file_extension)){
                 echo '<div class="plugin-name-' . $protectedGet['option'] . '">';
@@ -138,7 +150,7 @@ if (isset($protectedGet['cat']) && in_array($protectedGet['cat'], array('softwar
                 echo '<div class="plugin-name-' . $plugin->getId() . '">';
             }
         }
-        if(file_exists($file_extension)){
+        if(isset($file_extension) && file_exists($file_extension)){
             require $file_extension;
         }else{
             require $plugin_file;
